# IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "ECE 461 - Spring 2023 - Project 2",
    "description" : "API for ECE 461/Spring 2023/Project 2: A Trustworthy Module Registry",
    "termsOfService" : "http://swagger.io/terms/",
    "contact" : {
      "name" : "Prof. Davis",
      "url" : "http://davisjam.github.io",
      "email" : "davisjam@purdue.edu"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "2.0.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
    "/packages" : {
      "post" : {
        "summary" : "Get the packages from the registry.",
        "description" : "Get any packages fitting the query.\nSearch for packages satisfying the indicated query.\n\nIf you want to enumerate all packages, provide an array with a single PackageQuery whose name is \"*\".\n\nThe response is paginated; the response header includes the offset to use in the next query.",
        "operationId" : "PackagesList",
        "parameters" : [ {
          "name" : "X-Authorization",
          "in" : "header",
          "description" : "",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/AuthenticationToken"
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "description" : "Provide this for pagination. If not provided, returns the first page of results.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "$ref" : "#/components/schemas/EnumerateOffset"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/PackageQuery"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "List of packages",
            "headers" : {
              "offset" : {
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "$ref" : "#/components/schemas/EnumerateOffset"
                },
                "examples" : {
                  "ExampleRequest" : {
                    "value" : "3"
                  }
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PackageMetadata"
                  },
                  "x-content-type" : "application/json"
                },
                "examples" : {
                  "ExampleResponse" : {
                    "value" : [ {
                      "Version" : "1.2.3",
                      "Name" : "Underscore"
                    }, {
                      "Version" : "1.2.3-2.1.0",
                      "Name" : "Lodash"
                    }, {
                      "Version" : "^1.2.3",
                      "Name" : "React"
                    } ]
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "There is missing field(s) in the PackageQuery/AuthenticationToken or it is formed improperly, or the AuthenticationToken is invalid."
          },
          "413" : {
            "description" : "Too many packages returned."
          },
          "default" : {
            "description" : "unexpected error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/reset" : {
      "delete" : {
        "summary" : "Reset the registry",
        "description" : "Reset the registry to a system default state.",
        "operationId" : "RegistryReset",
        "parameters" : [ {
          "name" : "X-Authorization",
          "in" : "header",
          "description" : "",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/AuthenticationToken"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Registry is reset."
          },
          "400" : {
            "description" : "There is missing field(s) in the AuthenticationToken or it is formed improperly, or the AuthenticationToken is invalid."
          },
          "401" : {
            "description" : "You do not have permission to reset the registry."
          }
        }
      }
    },
...
    "/package" : {
      "post" : {
        "operationId" : "PackageCreate",
        "parameters" : [ {
          "name" : "X-Authorization",
          "in" : "header",
          "description" : "",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/AuthenticationToken"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PackageData"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Success. Check the ID in the returned metadata for the official ID.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Package"
                },
                "examples" : {
                  "ExampleResponse" : {
                    "value" : {
                      "metadata" : {
                        "Name" : "Underscore",
                        "Version" : "1.0.0",
                        "ID" : "underscore"
                      },
                      "data" : {
                        "Content" : "UEsDBBQAAAAAAA9DQlMAAAAAAAAAAAAAAAALACAAZXhjZXB0aW9ucy9VVA0AB35PWGF+T1hhfk9YYXV4CwABBPcBAAAEFAAAAFBLAwQUAAgACACqMCJTAAAAAAAAAABNAQAAJAAgAGV4Y2VwdGlvbnMvQ29tbWNvdXJpZXJFeGNlcHRpb24uamF2YVVUDQAH4KEwYeGhMGHgoTBhdXgLAAEE9wEAAAQUAAAAdY7NCoMwDMfvfYoct0tfQAYDGbv7BrVmW9DaksQhDN99BSc65gKBwP/jl+R86+4IPgabN/g4MCFbHD0mpdhLYQyFFFl/PIyijpVuzqvYCiVlO5axwWKJdDHUsbVXVEXOTef5MmmoO/LgOycC5dp5WbCAo2LfCFRDrxRwFV7GQJ7E9HSKsMUCf/0w+2bSHuPwN3vMFPiMPkjsVoTTHmcyk3kDUEsHCOEX4+uiAAAATQEAAFBLAwQUAAgACACqMCJTAAAAAAAAAAB9AgAAKgAgAGV4Y2VwdGlvbnMvQ29tbWNvdXJpZXJFeGNlcHRpb25NYXBwZXIuamF2YVVUDQAH4KEwYeGhMGHgoTBhdXgLAAEE9wEAAAQUAAAAdVHNTsMwDL7nKXzcJOQXKKCJwYEDAiHxACY1U0bbRI7bVUJ7d7JCtrbbIkVx4u/HdgLZb9owWF9j2rX1rTgW5N5yUOebWBjj6uBFzzDCUUnUfZHViA8U+Z1jSBQurlFadZVTxxEz9CO9jDy21FGPrtmyVXwejmKa20WUmESF8cxujOBe8Sl38UIhsFzFvYnvXHkAmFWOTWg/K2fBVhQjrE9NzEQhaVZcc6MRZqnbS6x7+DEG0lr9tTfEk2mAzGYzoF87FkmFDbf/2jIN1OdwcckTuF9m28Ma/9XRDe6g4d0kt1gWJ5KwttJMi8M2lKRH/CMpLTLgJrnihjUn175Mgllxb/bmF1BLBwiV8DzjBgEAAH0CAABQSwMEFAAIAAgAD0NCUwAAAAAAAAAAGQMAACYAIABleGNlcHRpb25zL0dlbmVyaWNFeGNlcHRpb25NYXBwZXIuamF2YVVUDQAHfk9YYX9PWGF+T1hhdXgLAAEE9wEAAAQUAAAAjVNRa8IwEH7Prwg+VZA87a3bcJsyBhNHx9hzTE+Npk25XG3Z8L8v7ZbaKsICaS6977vvu6QtpNrLDXBlM+FnpmyJGlBAraAgbXMXM6azwiJdYBAcSSS9loqceJQOEnCFp0D8P0qAP9n0OqUkbTRpOME//JuerZ08yFrofAeKxEu7xMNc5QQ6XxRBXDjsI6AmMQ+NL2RRAF7FvaE96LQHMDZb2X2TA8yFM+ubnXhvnt7ptA3YNJBYUa6MVlwZ6Rx/hhxQqzNl7usayCAnx89St93+nn8zxv2Y/jbexoNz4nh2ai16eQBE76Td/ZkJNE42hFEnxKEeB61m9G+7k+B3PIdqkIvG8Ylk7EZ4XYvR6KGpGGpX0nHaoq3y0aQR6lEQqMR82IQoi1RSJzGTJD81bWfgFOq2YhTwE97/xsQ8SZZJIyE2QK9WSaO/IF2Ac/4fiMZB+MiO7AdQSwcIIu3xZlgBAAAZAwAAUEsBAhQDFAAAAAAAD0NCUwAAAAAAAAAAAAAAAAsAIAAAAAAAAAAAAO1BAAAAAGV4Y2VwdGlvbnMvVVQNAAd+T1hhfk9YYX5PWGF1eAsAAQT3AQAABBQAAABQSwECFAMUAAgACACqMCJT4Rfj66IAAABNAQAAJAAgAAAAAAAAAAAApIFJAAAAZXhjZXB0aW9ucy9Db21tY291cmllckV4Y2VwdGlvbi5qYXZhVVQNAAfgoTBh4aEwYeChMGF1eAsAAQT3AQAABBQAAABQSwECFAMUAAgACACqMCJTlfA84wYBAAB9AgAAKgAgAAAAAAAAAAAApIFdAQAAZXhjZXB0aW9ucy9Db21tY291cmllckV4Y2VwdGlvbk1hcHBlci5qYXZhVVQNAAfgoTBh4aEwYeChMGF1eAsAAQT3AQAABBQAAABQSwECFAMUAAgACAAPQ0JTIu3xZlgBAAAZAwAAJgAgAAAAAAAAAAAApIHbAgAAZXhjZXB0aW9ucy9HZW5lcmljRXhjZXB0aW9uTWFwcGVyLmphdmFVVA0AB35PWGF/T1hhfk9YYXV4CwABBPcBAAAEFAAAAFBLBQYAAAAABAAEALcBAACnBAAAAAA=",
                        "JSProgram" : "if (process.argv.length === 7) {\nconsole.log('Success')\nprocess.exit(0)\n} else {\nconsole.log('Failed')\nprocess.exit(1)\n}\n"
                      }
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "There is missing field(s) in the PackageData/AuthenticationToken or it is formed improperly, or the AuthenticationToken is invalid."
          },
          "409" : {
            "description" : "Package exists already."
          },
          "424" : {
            "description" : "Package is not uploaded due to the disqualified rating."
          }
        }
      }
    },
    "/package/{id}/rate" : {
      "get" : {
        "operationId" : "PackageRate",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/PackageID"
          }
        }, {
          "name" : "X-Authorization",
          "in" : "header",
          "description" : "",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/AuthenticationToken"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Return the rating. Only use this if each metric was computed successfully.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PackageRating"
                }
              }
            }
          },
          "400" : {
            "description" : "There is missing field(s) in the PackageID/AuthenticationToken or it is formed improperly, or the AuthenticationToken is invalid."
          },
          "404" : {
            "description" : "Package does not exist."
          },
          "500" : {
            "description" : "The package rating system choked on at least one of the metrics."
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Error" : {
        "required" : [ "code", "message" ],
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "PackageID" : {
        "type" : "string",
        "description" : ""
      },
...
      "AuthenticationToken" : {
        "type" : "string",
        "description" : "The spec permits you to use any token format you like. You could, for example, look into JSON Web Tokens (\"JWT\", pronounced \"jots\"): https://jwt.io."
      },
...
      "SemverRange" : {
        "type" : "string",
        "description" : "",
        "example" : "Exact (1.2.3)\nBounded range (1.2.3-2.1.0)\nCarat (^1.2.3)\nTilde (~1.2.0)"
      },
      "PackageQuery" : {
        "required" : [ "Name" ],
        "type" : "object",
        "properties" : {
          "Version" : {
            "$ref" : "#/components/schemas/SemverRange"
          },
          "Name" : {
            "$ref" : "#/components/schemas/PackageName"
          }
        },
        "description" : ""
      },
      "EnumerateOffset" : {
        "type" : "string",
        "description" : "Offset in pagination.",
        "example" : "1"
      },
...
    },
    "securitySchemes" : {
      "UserAndPasswordYieldsToken" : {
        "type" : "apiKey",
        "description" : "Use the /authentication endpoint to obtain an authorization token.",
        "name" : "X-Authorization",
        "in" : "header"
      }
    }
  }
}